requirements.txt
python-docx
openai
google-generativeai
groq
requests